+++ spec/json/merge_spec.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "spec_helper"
+
 RSpec.describe Json::Merge do
   it "has a version number" do
     expect(Json::Merge::VERSION).not_to be_nil