+++ spec/support/shared_examples/file_analysis_examples.rb
@@ -481,4 +481,3 @@ RSpec.shared_examples "integrate_nodes" do
     end
   end
 end
-